requirements.txt
# tensorflow>=2.13.0
# tflite-runtime>=2.5.0
# Pillow==8.2.0
# Flask>=2.0.1
# gunicorn>=20.1.0
# Pillow>=8.2.0
# twilio>=6.59.0
# requests>=2.25.1
Flask>=2.0.1
gunicorn>=20.1.0
Pillow>=8.2.0
twilio>=6.59.0
requests>=2.25.1
numpy>=1.19.5
https://dl.google.com/coral/python/tflite_runtime-2.5.0-cp37-cp37m-linux_x86_64.whl
